# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Thrive API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /plans:
    get:
      operationId: listTrainingPlans
      tags: [plans]
      summary: List training plans
      responses:
        "200":
          description: List of training plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrainingPlan"
    post:
      operationId: createTrainingPlan
      tags: [plans]
      summary: Create a training plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainingPlanInput"
      responses:
        "201":
          description: Created plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingPlan"
  /plans/{id}:
    get:
      operationId: getTrainingPlan
      tags: [plans]
      summary: Get a training plan with sessions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Training plan with sessions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingPlanDetail"
        "404":
          description: Not found
    patch:
      operationId: updateTrainingPlan
      tags: [plans]
      summary: Update a training plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainingPlanUpdate"
      responses:
        "200":
          description: Updated training plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingPlanDetail"
    delete:
      operationId: deleteTrainingPlan
      tags: [plans]
      summary: Delete a training plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /plans/{planId}/sessions/{sessionId}:
    patch:
      operationId: updatePlanSession
      tags: [plans]
      summary: Update a plan session
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanSessionUpdate"
      responses:
        "200":
          description: Updated plan session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanSession"
...
components:
...
  schemas:
...
    TrainingPlan:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        goal:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        weeklyMileage:
          type: ["number", "null"]
        status:
          type: string
          enum: [active, completed, paused]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - goal
        - startDate
        - endDate
        - status
        - createdAt
    TrainingPlanDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        goal:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        weeklyMileage:
          type: ["number", "null"]
        status:
          type: string
          enum: [active, completed, paused]
        createdAt:
          type: string
          format: date-time
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/PlanSession"
      required:
        - id
        - name
        - goal
        - startDate
        - endDate
        - status
        - createdAt
        - sessions
    PlanSession:
      type: object
      properties:
        id:
          type: integer
        planId:
          type: integer
        weekNumber:
          type: integer
        dayOfWeek:
          type: integer
        sessionType:
          type: string
        description:
          type: string
        distanceKm:
          type: ["number", "null"]
        durationMinutes:
          type: ["integer", "null"]
        completed:
          type: boolean
      required:
        - id
        - planId
        - weekNumber
        - dayOfWeek
        - sessionType
        - description
        - completed
    TrainingPlanInput:
      type: object
      properties:
        name:
          type: string
        goal:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        weeklyMileage:
          type: number
      required:
        - name
    TrainingPlanUpdate:
      type: object
      properties:
        name:
          type: string
        goal:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        weeklyMileage:
          type: ["number", "null"]
        status:
          type: string
          enum: [active, completed, paused]
    PlanSessionUpdate:
      type: object
      properties:
        completed:
          type: boolean
        description:
          type: string
        distanceKm:
          type: ["number", "null"]
        durationMinutes:
          type: ["integer", "null"]
...